Move videos to a new local videos page on Library screen
