+++ lib/pages/components/localVideos/videosOnFolderListView.dart
@@ -8,20 +8,16 @@ import 'package:songtube/internal/languages.dart';
 
 // Internal
 import 'package:songtube/internal/models/videoFile.dart';
-import 'package:songtube/ui/animations/showUp.dart';
 
 // Packages
-import 'package:eva_icons_flutter/eva_icons_flutter.dart';
 import 'package:transparent_image/transparent_image.dart';
 
 class VideosOnFolderListView extends StatelessWidget {
   final List<VideoFile> list;
   final Function(VideoFile) onVideoTap;
-  final Function onBackPressed;
   VideosOnFolderListView({
     @required this.list,
     @required this.onVideoTap,
-    @required this.onBackPressed
   });
   @override
   Widget build(BuildContext context) {
@@ -56,11 +52,12 @@ class VideosOnFolderListView extends StatelessWidget {
                             return ClipRRect(
                               borderRadius: BorderRadius.circular(10),
                               child: Stack(
+                                fit: StackFit.expand,
                                 children: [
                                   FadeInImage(
                                     fadeInDuration: Duration(milliseconds: 300),
                                     placeholder: MemoryImage(kTransparentImage),
-                                    image: thumbnail.hasData && thumbnail.data.existsSync()
+                                    image: thumbnail.hasData && thumbnail.data != null
                                       ? FileImage(thumbnail.data)
                                       : MemoryImage(kTransparentImage),
                                     fit: BoxFit.cover,
@@ -79,11 +76,13 @@ class VideosOnFolderListView extends StatelessWidget {
                                         builder: (context, AsyncSnapshot<int> videoDuration) {
                                           return Text(
                                             videoDuration.hasData
-                                              ? "${Duration(milliseconds: videoDuration.data).inMinutes}min"
+                                              ? "${Duration(seconds: videoDuration.data).inMinutes}min"
                                               : Languages.of(context).labelCalculating,
                                             style: TextStyle(
                                               color: Colors.white,
-                                              fontSize: 12
+                                              fontSize: 12,
+                                              fontFamily: 'Product Sans',
+                                              fontWeight: FontWeight.w600
                                             ),
                                           );
                                         },
@@ -109,7 +108,8 @@ class VideosOnFolderListView extends StatelessWidget {
                         .replaceAll(".mkv", ''),
                       style: TextStyle(
                         fontSize: 18,
-                        fontWeight: FontWeight.w500
+                        fontFamily: 'Product Sans',
+                        fontWeight: FontWeight.w600
                       ),
                       maxLines: 2,
                       textAlign: TextAlign.start,
@@ -122,7 +122,9 @@ class VideosOnFolderListView extends StatelessWidget {
                       "${(int.parse(video.size)/1000000).toStringAsFixed(2)}MB",
                       style: TextStyle(
                         fontSize: 12,
-                        color: Theme.of(context).textTheme.bodyText1.color.withOpacity(0.6)
+                        color: Theme.of(context).textTheme.bodyText1.color.withOpacity(0.6),
+                        fontFamily: 'Product Sans',
+                        fontWeight: FontWeight.w600
                       ),
                       maxLines: 1,
                       textAlign: TextAlign.start,
@@ -133,17 +135,6 @@ class VideosOnFolderListView extends StatelessWidget {
             ),
           );
         },
-      ),
-      floatingActionButton: ShowUpTransition(
-        duration: Duration(milliseconds: 300),
-        delay: Duration(milliseconds: 100),
-        forward: true,
-        slideSide: SlideFromSlide.BOTTOM,
-        child: FloatingActionButton(
-          onPressed: onBackPressed,
-          child: Icon(EvaIcons.arrowBackOutline, color: Colors.white),
-          backgroundColor: Theme.of(context).accentColor,
-        ),
       ),
     );
   }
